pyproject.toml
[project]
name = "silo"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "authlib>=1.6.3",
    "cryptography>=45.0.7",
    "django>=5.2.6",
    "djangorestframework>=3.16.1",
    "python-dotenv>=1.1.1",
    "python-jose>=3.5.0",
    "requests>=2.32.5",
    "ruff>=0.13.0",
]
